Refactor for code clarity
