+++ packages/web-app/app/_ui/components/acquire-access-grant-button.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { Button } from '.';
-import { PublicInfo } from '@/app/server/types';
+import { PublicInfo } from '@/app/_server/types';
 import { useIdOS } from '@/app/_providers/idos';
 
 type AcquireAccessGrantButton = {
