+++ packages/web-app/app/_ui/components/index.tsx
@@ -1,5 +1,5 @@
 export { Button } from './button';
-export { Info } from './info';
+export { Info } from '../projects/info';
 export { Topbar } from './topbar';
 export { Footer } from './footer';
 export { WalletButton } from './wallet-button';
