+++ packages/web-app/app/_ui/projects/info-with-server-details.tsx
@@ -1,0 +1,8 @@
+import { getPublicInfo } from '@/app/_server/idos';
+import { Info } from './info';
+
+export async function InfoWithServerDetails() {
+  const serverInfo = await getPublicInfo();
+
+  return <Info serverInfo={serverInfo} />;
+}
