+++ packages/web-app/app/projects/[id]/page.tsx
@@ -1,14 +1,36 @@
-import { getPublicInfo } from '@/app/server/idos';
-import { Content } from './content';
+import { InfoWithServerDetails } from '@/app/_ui/projects/info-with-server-details';
+import { ProjectDescription } from '@/app/_ui/projects/project-description';
+import { TokenMetrics } from '@/app/_ui/projects/token-metrics';
+import { Suspense } from 'react';
 
-type TPageProps = {
+type TProjectProps = {
   params: {
     id: string;
   };
 };
 
-export default async function Page(props: TPageProps) {
-  const serverInfo = await getPublicInfo();
+export default function Project({ params }: TProjectProps) {
+  if (params.id !== 'citizend') {
+    return '404 Not Found';
+  }
 
-  return <Content serverInfo={serverInfo} {...props} />;
+  return (
+    <main className="grid grid-cols-1 gap-x-17 gap-y-9 px-31 lg:grid-cols-2">
+      <ProjectDescription />
+      <div className="lg:pt-27">
+        <TokenMetrics
+          token={'RDF'}
+          targetRaiseRange="2.M - 8M ETH"
+          totalSupply="800 000 000 RDF"
+          targetedDate={new Date('2024-04-29')}
+          contractAddress="0x02323430947"
+        />
+        <div className="mt-9">
+          <Suspense key={params.id} fallback={<div>Loading...</div>}>
+            <InfoWithServerDetails />
+          </Suspense>
+        </div>
+      </div>
+    </main>
+  );
 }
